refactor: refactor app architecture
