+++ src/components/elements/localeSwitcher/switcher.module.scss
