+++ src/components/pages/welcomePage/components/author-card/AuthorCard.tsx
