+++ src/components/pages/welcomePage/components/author-card/author-card.module.scss
