+++ src/components/pages/welcomePage/components/author-card/author-text/Author-text.tsx
