+++ src/components/pages/welcomePage/components/author-card/authorsInfo.ts
