+++ src/components/pages/welcomePage/components/authors-slider/AuthorsSlider.tsx
