+++ src/components/pages/welcomePage/components/authors-slider/authors-slider.module.scss
