+++ src/components/pages/welcomePage/components/authorsSection/AuthorsSection.tsx
