+++ src/components/pages/welcomePage/components/authorsSection/authorsSection.module.scss
