+++ src/tests/app/locale/components/header/header.test.tsx
@@ -1,4 +1,4 @@
-import LocaleSwitcherButton from "@/components/elements/LocaleSwitcherButton";
+import LocaleSwitcherButton from "@/components/elements/localeSwitcher/LocaleSwitcherButton";
 import Header from "@/components/header/Header";
 import { IntlProviderWrapper } from "@/tests/utils/test-utils";
 import { render, screen, fireEvent } from "@testing-library/react";
